requirements.txt
pillow==9.0.1
streamlit~=1.5.1
numpy~=1.19.5
plotly~=5.3.1
gsheetsdb~=0.1.13
gspread==5.1.1
gitdb~=4.0.9
defusedxml~=0.7.1
cffi~=1.15.0
future~=0.18.2
ipython~=8.0.1
pip~=21.1.2
attrs~=21.4.0
toml~=0.10.2
wheel~=0.36.2
google~=3.0.0
lxml~=4.7.1
tornado~=6.1
Jinja2~=3.0.3
rsa~=4.8
pyasn1~=0.4.8
toolz~=0.11.2
pyzmq~=22.3.0
colorama~=0.4.4
parso~=0.8.3
jedi~=0.18.1
pytz~=2021.3
astor~=0.8.1
smmap~=5.0.0
pywin32~=303
setuptools~=57.0.0
pydeck~=0.7.1
pandas~=1.3.5
ipykernel~=6.9.1
ipywidgets~=7.6.5
altair~=4.2.0
jsonschema~=4.4.0
entrypoints~=0.4
base58~=2.1.1
six~=1.16.0
webencodings~=0.5.1
packaging~=21.3
bleach~=4.1.0
cachetools~=4.2.4
certifi~=2021.10.8
urllib3~=1.26.8
requests~=2.27.1
MarkupSafe~=2.0.1
python-dateutil~=2.8.2
pyarrow~=7.0.0
tenacity~=8.0.1
traitlets~=5.1.1
blinker~=1.4
debugpy~=1.5.1
decorator~=5.1.1
Pygments~=2.11.2
testpath~=0.5.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.5
tzlocal~=4.1
nbformat~=5.1.3
nbclient~=0.5.11
nbconvert~=6.4.2
notebook~=6.4.8
Send2Trash~=1.8.0
terminado~=0.13.1
oauthlib~=3.1.1
idna~=3.3
watchdog~=2.1.6
pyparsing~=3.0.7
pandocfilters~=1.5.0
mistune~=0.8.4
pycparser~=2.21
click~=8.0.3
validators~=0.18.2
pyrsistent~=0.18.1